requirements.txt
python-telegram-bot>=20.0
yt-dlp>=2023.12.30
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiofiles>=23.2.1
psutil>=5.9.0
